+++ src/benchmarkstt/diff/__init__.py
@@ -2,16 +2,19 @@
 Responsible for calculating differences.
 """
 
-from benchmarkstt.factory import Factory
+from abc import ABC, abstractmethod
+from benchmarkstt.factory import CoreFactory
 
 
-class Differ:
+class Differ(ABC):
+    @abstractmethod
     def __init__(self, a, b):
         """
         :meta public:
         """
         raise NotImplementedError()
 
+    @abstractmethod
     def get_opcodes(self):
         """
         Return list of 5-tuples describing how to turn `a` into `b`.
@@ -30,4 +33,4 @@ class Differ:
         raise NotImplementedError()
 
 
-factory = Factory(Differ)
+factory = CoreFactory(Differ)