+++ src/benchmarkstt/metrics/__init__.py
@@ -3,16 +3,18 @@ Responsible for calculating metrics.
 
 """
 
+from abc import ABC, abstractmethod
 from benchmarkstt.schema import Schema
-from benchmarkstt.factory import Factory
+from benchmarkstt.factory import CoreFactory
 
 
-class Metric:
+class Metric(ABC):
     """
     Base class for metrics
     """
+    @abstractmethod
     def compare(self, ref: Schema, hyp: Schema):
         raise NotImplementedError()
 
 
-factory = Factory(Metric)
+factory = CoreFactory(Metric)