+++ src/benchmarkstt/output/__init__.py
@@ -2,11 +2,12 @@
 Responsible for dealing with output formats
 
 """
+from abc import ABC, abstractmethod
+from collections import OrderedDict
+from benchmarkstt.factory import CoreFactory
 
-from benchmarkstt.factory import Factory
 
-
-class Output:
+class Output(ABC):
     def __enter__(self):
         """
         :meta public:
@@ -19,6 +20,7 @@ class Output:
         """
         pass
 
+    @abstractmethod
     def result(self, title, result):
         """
         :meta public:
@@ -26,4 +28,26 @@ class Output:
         raise NotImplementedError()
 
 
-factory = Factory(Output)
+class SimpleTextBase(Output):
+    @staticmethod
+    def print(result):
+        if hasattr(result, '_asdict'):
+            result = result._asdict()
+
+        if type(result) is float:
+            print("%.6f" % (result,))
+        elif type(result) is dict or type(result) is OrderedDict:
+            for k, v in result.items():
+                print("%s: %r" % (k, v))
+        else:
+            print(result)
+
+    @abstractmethod
+    def result(self, title, result):
+        """
+        :meta public:
+        """
+        raise NotImplementedError()
+
+
+factory = CoreFactory(Output)