+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 
 
   * add auto-generated UML diagrams
+  * add tutorial Jupyter Notebooks
 
 * 
   Tests:
@@ -32,6 +33,10 @@ Changed
   * moved all documentation specific code outside main package
   * update sphinx to latest
   * use more descriptive names for Base classes (Normalizer, Differ, etc.)
+  * rename CLIDiffDialect to ANSIDiffDialect, "cli" -> "ansi"
+  * rename NormalizationComposite -> NormalizationAggregate
+  * allow ducktyped custom classes to be recognized as valid
+  * proper abstract base classes
 
 * 
   Documentation: